Fix import and reference errors - Fix import errors added through the Environment variables feature
