+++ tsviewer/configuration.py
@@ -2,6 +2,7 @@ from tsviewer.ts_viewer_utils import resolve_with_project_path
 from dataclasses import dataclass
 from json import load, dump
 import ts3.query
+from os import environ
 
 
 @dataclass
@@ -44,3 +45,24 @@ def authorize(configuration: Configuration, connection: ts3.query.TS3Connection)
     """
     connection.login(client_login_name=configuration.user, client_login_password=configuration.password)
     connection.use(sid=configuration.server_id)
+
+
+def read_environment_variables(configuration: Configuration) -> None:
+    """
+    Overwrite all configuration fields when there are configured environment variables for those fields.
+    The pattern for those variables is TSVIEWER_CONFIGURATION_FIELD_NAME. E.g.: TSVIEWER_CONFIGURATION_PASSWORD
+    :param configuration: configuration object
+    """
+    for name in configuration.__dict__.keys():
+        value = environ.get(f'TSVIEWER_CONFIGURATION_{name.upper()}')
+        if value:
+            configuration.__setattr__(name, value)
+
+
+def read_config_path_from_environment_variables(default_path: str = 'config/config.json') -> str:
+    """
+    Read the file path from the environment variable TSVIEWER_CONFIGURATION_FILE if configured
+    :param default_path:
+    :return:
+    """
+    return environ.get('TSVIEWER_CONFIGURATION_FILE', default_path)
